Fixed linker errors on Windows when turned on logging. Fixed crashing time formatting on Windows.
